Update master to 4.1.0
* git reset --hard

To ditch those 60 other commits.
Git please

* Bugsnag this shit up boi

* wew

when the bugsnag example produces TypeErrors

* wew x2

* wew yet again

* Let's not report to bugsnag on nothing

* Update help command with new command reference

* catch a crash when a user does join-voice, requests a playlist and makes the bot leave. Should no longer look the temp list and keep fetching.

* voice,request,leave-voice before it's added to the playlist leads to a crash, thix fixes the problem.

* playlist delete feature

* smh i'm an idiot, try to fix the typo with another one

* who thought catch would update the user?

* fix my last commit with the proper thing that should have been!

* because ya know.. make sure it's a song that can actually be deleted

* rethinkdb databes controllzerz (#185)

To fuck, or to clusterfuck, that is the question

* kebab fix (#186)

* delete or remove, word of choice

* Lets not spam our bugsnag shall we (#187)

* Lets not spam bugsnag

* Bugsnag key support

* Update voice.js

* Update default.js

* Update customize.js

* Update commandcontrol.js

* Update DougBot.js

* Update voice.js

* Update DougBot.js

* ok dad

* sigh

* GOD DAMN IT

* :^)

* wew

* @Nenkai pls

* well at least it's a fix, maybe someone can do it better?

* i forgot to reject the promise, oops.

* well at least it's a fix, maybe someone can do it better? (#190)

* well at least it's a fix, maybe someone can do it better?

* i forgot to reject the promise, oops.

* resolve not reject, screw the error message lol

* I hate git, i just, can't even sometimes.

* ayy lmao xkcd

* whoops forgot to use ===

* because nenkai makes me lose faith in people

* zezu pls (#191)

* so i finally did that `npm run dbcreate` thing i was talking about (#192)

* so i finally did that `npm run dbcreate` thing i was talking about

* i linted aand well this happened

* Remove version 2.1.1 or whatever remnants

People actually thought WildBot logged chats because of this, even when that export wasn't used anywhere.

* Fix dbcreate

* Blacklist the dbcreate script from being run by the bot, that should not happen. Delete the old method. (#194)

* Update Info command to now use embeds (#195)

* Updated Info command to use embeds

* Update Discordie, for embeds support

* Userinfo to use embeds (#197)

* Userinfo to use embeds 👌

* Almost forgot

* ...

* Serverinfo > Embed

* Serverinfo roles fix (#198)

* Update default.js (#199)

* Track guild owner changes.

* Tag check missing (#200)

* When nenkai breaks shit again 👀 (#201)

* (╯°□°）╯︵ ┻━┻

* Remove obsolete CSE field, customize help update

* *violently spams Ctrl+S*

* People complain when I forget this

* Implement rankup command (#203)

* Implement rankup command

* Logic changes

* Avoid multiple messages

* Doug, you forgot something

@Dougley I will still complain because you forgot a command.

* Doug, you forgot something

@SteamingMutt I will still complain because you forgot a command.

* Docker should create tables

* New npm who dis

* Slightly rewrite README for 4.1.0 release

* Release preparatory changelog edit

+++ runtime/commands/music.js
@@ -1,5 +1,6 @@
 'use strict'
 var v = require('../internal/voice.js')
+var checkLevel = require('../databases/controllers/permissions.js').checkLevel
 var Commands = []
 
 Commands.music = {
@@ -66,30 +67,54 @@ Commands.skip = {
 
 Commands.playlist = {
   name: 'playlist',
-  help: "I'll fetch you the playlist I'm currently playing!",
+  help: "Use delete and a song number to remove it from the list else I will fetch you the playlist I'm currently playing!",
   aliases: ['list'],
   noDM: true,
-  timeout: 10,
+  timeout: 5,
   level: 0,
-  fn: function (msg) {
-    v.fetchList(msg).then((r) => {
-      var arr = []
-      arr.push('Now playing: **' + r.info[0] + '** \n')
-      for (var i = 1; i < r.info.length; i++) {
-        arr.push((i + 1) + '. **' + r.info[i] + '** Requested by ' + r.requester[i])
-        if (i === 9) {
-          arr.push('And about ' + (r.info.length - 10) + ' more songs.')
-          break
-        }
+  fn: function (msg, suffix, bot) {
+    suffix = suffix.toLowerCase().split(' ')
+    var connect = bot.VoiceConnections.find(v => v.voiceConnection.guild.id === msg.guild.id)
+    if (connect) {
+      if ((suffix[0] === 'delete' || suffix[0] === 'remove') && suffix[1] >= 2) {
+        checkLevel(msg, msg.author.id, msg.member.roles).then(function (r) {
+          if (r >= 1) {
+            v.deleteFromPlaylist(msg, suffix[1] - 1).then(s => {
+              msg.channel.sendMessage(`**${s}** has been removed from the playlist`).then(m => {
+                setTimeout(() => {
+                  m.delete()
+                }, 15000)
+              })
+            }).catch(e => {
+              msg.channel.sendMessage(e)
+            })
+          } else {
+            msg.channel.sendMessage('You have no permission to run this command!\nYou need level 1, you have level ' + r + '\nAsk the server owner to modify your level with `setlevel`.')
+          }
+        })
+      } else {
+        v.fetchList(msg).then((r) => {
+          var arr = []
+          arr.push('Now playing: **' + r.info[0] + '** \n')
+          for (var i = 1; i < r.info.length; i++) {
+            arr.push((i + 1) + '. **' + r.info[i] + '** Requested by ' + r.requester[i])
+            if (i === 9) {
+              if (r.info.length - 10 !== 0) arr.push('And about ' + (r.info.length - 10) + ' more songs.')
+              break
+            }
+          }
+          msg.channel.sendMessage(arr.join('\n')).then((m) => {
+            setTimeout(() => {
+              m.delete()
+            }, 15000)
+          })
+        }).catch(() => {
+          msg.channel.sendMessage("It appears that there aren't any songs in the current queue.")
+        })
       }
-      msg.channel.sendMessage(arr.join('\n')).then((m) => {
-        setTimeout(() => {
-          m.delete()
-        }, 15000)
-      })
-    }).catch(() => {
-      msg.channel.sendMessage("It appears that there aren't any songs in the current queue.")
-    })
+    } else {
+      msg.channel.sendMessage('I am not streaming music in this server.')
+    }
   }
 }
